Unified ItemContentDetails with tvOS.
