+++ Shared/View/Details/InfoSegmentView.swift
@@ -1,0 +1,32 @@
+//
+//  InfoSegmentView.swift
+//  Story (iOS)
+//
+//  Created by Alexandre Madeira on 05/05/23.
+//
+
+import SwiftUI
+
+struct InfoSegmentView: View {
+    let title: String
+    let info: String?
+    var body: some View {
+        if let info {
+            VStack(alignment: .leading) {
+                Text(NSLocalizedString(title, comment: ""))
+                    .lineLimit(1)
+                    .font(.body)
+                    .foregroundColor(.secondary)
+                Text(info)
+                    .lineLimit(1)
+                    .font(.body)
+            }
+        }
+    }
+}
+
+struct InfoSegmentView_Previews: PreviewProvider {
+    static var previews: some View {
+        InfoSegmentView(title: "This is for Preview", info: "SwiftUI Previews!")
+    }
+}
